splitting train_utils
